Use absolute path of grab2.py for the kitten call
When it spawns a new kitten, it would do so calling the grab2.py
script directly. That would work when both files were in the same
directory, and they were on ~/.config/kitty/. which kitty apparently
runs/spawns kittens from. However, as soon as any of those two rules
didn't meet, this kitten would fail.

So instead of hard-coding the path of grab2.py, we're making use of
Python's import capabilities. A directory is considered a package or a
namespace. Either way, it means that no matter where the files are
being imported from, a file from a package can call and access files
from its same package. Then, using the `__file__` attribute of the
imported `grab2` module, and `os.path.abspath`, we can get its
absolute path in the system, which is exactly what I need for the
script not to fail no matter in what folder do I decide to save my
kittens.